fix(e2e): reconcile with PAY-38's deposits table readability change
Rebasing onto origin/main (PR 122, PAY-38) surfaced a real collision that
neither change would have caught alone.

PAY-38 changed the deposits period column to a three-letter month —
`periodLabel` now renders "Oct 2025", not "October 2025". Two consequences:

1. PAY-9 built its expected label with `month: "long"`, so it would have gone
   red on main the moment this branch landed. Switched to `month: "short"`.
2. PAY-38's own new assertions in journey 7 read `.p-datatable-tbody tr`
   `.first()` and expect it to be the October row. That holds today only
   because the boot's 2025-10 run is the sole 2025 deposit; this branch seeds a
   full year for three personas, so the first row becomes December's and those
   assertions would have started testing the wrong record. All of them are now
   scoped to the October row explicitly.

Kept every PAY-38 assertion — three-letter month, no EFTPS string in the table
body, sortable columns present — just anchored to a named row instead of a
positional one.

Verified after rebase: ephemeral e2e 20 passed / 1 skipped / 0 failed; lint 0
errors / 539 warnings = baseline; build; typecheck; 645 unit+integration tests.

+++ e2e/tests/qa.spec.ts
@@ -183,7 +183,9 @@ test("tax deposits: admin sees the computed schedule incl. last month (PAY-9)", 
     // has produced no current-year runs yet — the default view is empty.
     await page.goto(`/admin/deposits?year=${prev.getUTCFullYear()}`);
     await expect(page.getByRole("heading", { name: "Tax deposits" })).toBeVisible();
-    const label = `${prev.toLocaleString("en-US", { month: "long", timeZone: "UTC" })} ${prev.getUTCFullYear()}`;
+    // Three-letter month, matching AdminDepositsView's periodLabel since PAY-38
+    // ("Aug 2026", not "August 2026"). `month: "short"` gives the same list.
+    const label = `${prev.toLocaleString("en-US", { month: "short", timeZone: "UTC" })} ${prev.getUTCFullYear()}`;
     await expect(page.locator("tbody tr", { hasText: label }).first()).toBeVisible();
 
     // Jurisdiction + reminder schedule editor render.
